+++ client.go
@@ -1348,6 +1348,18 @@ type ReqUploadMedia struct {
 	UploadURL string
 }
 
+func (cli *Client) tryUploadMediaToURL(url, contentType string, content io.Reader) (*http.Response, error) {
+	cli.Logger.Debugfln("Uploading media to external URL %s", url)
+	req, err := http.NewRequest(http.MethodPut, url, content)
+	if err != nil {
+		return nil, err
+	}
+	req.Header.Set("Content-Type", contentType)
+	req.Header.Set("User-Agent", cli.UserAgent+" (external media uploader)")
+
+	return http.DefaultClient.Do(req)
+}
+
 func (cli *Client) uploadMediaToURL(data ReqUploadMedia) (*RespMediaUpload, error) {
 	retries := cli.DefaultHTTPRetries
 	if data.ContentBytes == nil {
@@ -1355,29 +1367,26 @@ func (cli *Client) uploadMediaToURL(data ReqUploadMedia) (*RespMediaUpload, erro
 		retries = 0
 	}
 	for {
-		if data.Content == nil {
-			data.Content = bytes.NewReader(data.ContentBytes)
+		reader := data.Content
+		if reader == nil {
+			reader = bytes.NewReader(data.ContentBytes)
+		} else {
+			data.Content = nil
 		}
-		cli.Logger.Debugfln("Uploading media to external URL %s", data.UploadURL)
-		req, err := http.NewRequest(http.MethodPut, data.UploadURL, data.Content)
-		if err != nil {
+		resp, err := cli.tryUploadMediaToURL(data.UploadURL, data.ContentType, reader)
+		if err == nil {
+			if resp.StatusCode >= 200 && resp.StatusCode < 300 {
+				// Everything is fine
+				break
+			}
+			err = fmt.Errorf("HTTP %d", resp.StatusCode)
+		}
+		if retries <= 0 {
+			cli.logWarning("Error uploading media to %s: %v, not retrying", data.UploadURL, err)
 			return nil, err
 		}
-		// Tell the next retry to create a new reader from ContentBytes
-		data.Content = nil
-		req.Header.Set("Content-Type", data.ContentType)
-		req.Header.Set("User-Agent", cli.UserAgent+" (external media uploader)")
-
-		resp, err := http.DefaultClient.Do(req)
-		if err != nil {
-			cli.Logger.Debugfln("Error uploading media to %s: %v, retrying", data.UploadURL, err)
-			retries--
-		} else if resp.StatusCode >= 400 || resp.StatusCode < 200 {
-			cli.Logger.Debugfln("Error uploading media to %s: HTTP %d, retrying", data.UploadURL, resp.StatusCode)
-			retries--
-		} else {
-			break
-		}
+		cli.Logger.Debugfln("Error uploading media to %s: %v, retrying", data.UploadURL, err)
+		retries--
 	}
 
 	query := map[string]string{}
@@ -1503,7 +1512,7 @@ func (cli *Client) Hierarchy(roomID id.RoomID, req *ReqHierarchy) (resp *RespHie
 // Messages returns a list of message and state events for a room. It uses
 // pagination query parameters to paginate history in the room.
 // See https://spec.matrix.org/v1.2/client-server-api/#get_matrixclientv3roomsroomidmessages
-func (cli *Client) Messages(roomID id.RoomID, from, to string, dir rune, filter *FilterPart, limit int) (resp *RespMessages, err error) {
+func (cli *Client) Messages(roomID id.RoomID, from, to string, dir Direction, filter *FilterPart, limit int) (resp *RespMessages, err error) {
 	query := map[string]string{
 		"from": from,
 		"dir":  string(dir),
@@ -1524,6 +1533,19 @@ func (cli *Client) Messages(roomID id.RoomID, from, to string, dir rune, filter 
 
 	urlPath := cli.BuildURLWithQuery(ClientURLPath{"v3", "rooms", roomID, "messages"}, query)
 	_, err = cli.MakeRequest("GET", urlPath, nil, &resp)
+	return
+}
+
+// TimestampToEvent finds the ID of the event closest to the given timestamp.
+//
+// See https://spec.matrix.org/v1.6/client-server-api/#get_matrixclientv1roomsroomidtimestamp_to_event
+func (cli *Client) TimestampToEvent(roomID id.RoomID, timestamp time.Time, dir Direction) (resp *RespTimestampToEvent, err error) {
+	query := map[string]string{
+		"ts":  strconv.FormatInt(timestamp.UnixMilli(), 10),
+		"dir": string(dir),
+	}
+	urlPath := cli.BuildURLWithQuery(ClientURLPath{"v1", "rooms", roomID, "timestamp_to_event"}, query)
+	_, err = cli.MakeRequest(http.MethodGet, urlPath, nil, &resp)
 	return
 }
 
@@ -1820,6 +1842,12 @@ func (cli *Client) BeeperMergeRooms(req *ReqBeeperMergeRoom) (resp *RespBeeperMe
 func (cli *Client) BeeperSplitRoom(req *ReqBeeperSplitRoom) (resp *RespBeeperSplitRoom, err error) {
 	urlPath := cli.BuildClientURL("unstable", "com.beeper.chatmerging", "rooms", req.RoomID, "split")
 	_, err = cli.MakeRequest(http.MethodPost, urlPath, req, &resp)
+	return
+}
+
+func (cli *Client) BeeperDeleteRoom(roomID id.RoomID) (err error) {
+	urlPath := cli.BuildClientURL("unstable", "com.beeper.yeet", "rooms", roomID, "delete")
+	_, err = cli.MakeRequest(http.MethodPost, urlPath, nil, nil)
 	return
 }